requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.6
cffi==1.12.1
Click==7.0
cryptography==2.5
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
junos-eznc==2.2.0
lxml==4.3.1
MarkupSafe==1.1.0
ncclient==0.6.3
netaddr==0.7.19
paramiko==2.4.2
pkg-resources==0.0.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyNaCl==1.3.0
pyserial==3.4
python-ldap==3.1.0
PyYAML==3.13
scp==0.13.0
selectors2==2.0.1
six==1.12.0
Werkzeug==0.14.1
